mini_gpt/requirements.txt
openai
flask
python-dotenv
mysql-connector-python
flask_login

